ostree: cleanly shut down HTTP server
When not calling server.shutdown(), one got errors like this:
  Exception in thread OSTree HTTPD:
  ValueError: file descriptor cannot be a negative integer (-1)

Probably caused by closing the socket while the server listening to it
was still active.

Worse, with Python 3.5.3 (Debian Stretch), the code now hangs during
shutdown of oe-selftest, with two threats waiting in mutex operations.

This does not happen when calling server.shutdown(), so let's call it
and debug whatever problems it might cause with other Python versions
when it fails again.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>