refkit distro: remove broken group ID entries
"dbus" and "dbus-x11" were added to Ostro because "they seem to be
used only during the build". However, the format was wrong (two fields
instead of three). Either they were not really needed in the first
place or are not used anymore. We better remove them before sorting by
numeric ID.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>