secureboot.py: avoid side effects
The test needs to build an image multiple times with different image
settings and cleans it when done. Using the normal refkit-image-common
affects the time required for other tests (like the installer test,
which can use an existing, unmodified refkit-image-common) and
developers (refkit-image-common is gone after the test).

Using an image recipe that is identical to refkit-image-common but
has a different name avoids this.

[YOCTO #11757]

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>