secureboot.py: ensure that efivars is installed
secureboot_with_image calls efivars regardless whether signing is
enabled or not, and therefore it always has to be installed.

Not sure how it worked before, perhaps because efivars was pulled in
for other reasons. But after the recent changes,
secureboot.SecureBootTests.test_boot_unsigned_image started to fail
with:

  File ".../cases/secureboot.py", line 129, in secureboot_with_image
    self.assertTrue(status, 'SecureBoot not enabled (status=%s):\n%s' % (status, output))
  File "/usr/lib/python3.5/unittest/case.py", line 678, in assertTrue
    raise self.failureException(msg)
AssertionError: 0 is not true : SecureBoot not enabled (status=0):
-sh: efivar: not found
-sh: efivar: not found

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>