polkit_%.bbappend: polkit client library + groupcheck
We want to enable code which relies on polkit for authorization
checks (like fwupd), but we don't want to pull the actual polkit
server plus mozjs into the build.

So here we simplify the build by patching out the service and the
mozjs dependency check, then package the shared client library with a
runtime dependency on groupcheck.

The newer release of groupcheck with support for reading per-package
policy files is used and configured so that /usr/share/groupcheck.d
is checked for policy files.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>