CI: build also x11 and wayland images
Regressions have not been noticed unless someone manually built images
with graphics. We need to do at least build testing, and ideally
produce images that then can also be tested with DAFT.

We could build the normal configuration and the configs where X11 and
Wayland are enabled in parallel, but then due to
https://bugzilla.yoctoproject.org/show_bug.cgi?id=10682 we'd end up
building the same base tools multiple times when they need to be
rebuilt.

So instead, here we do a normal initial build to populate the sstate
cache, then do a multiconfig build for x11 and wayland in parallel.

Unique image recipes are used merely to simplify the image handling
once the build is done. We could also build, say,
multiconfig:x11:refkit-image-common, and then rename
tmp-glibc-x11/deploy/images/*/refkit-image-common-intel-corei7-64.wic.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>