CI: call default config "headless"
Now that we have "x11" and "wayland" builds, it looks odd to not have
a name for the default configuration. For example, the tmp-glibc
directory would end up getting published as just "glibc".

It's better to explicitly use "tmp-headless-glibc", which then
produces "headless-glibc" alongside "x11-glibc" and
"wayland-glibc". "headless" is better than "default" or something like
that because it describes the content instead of assuming that people
know what the default is.

If we want, we can then drop "glibc" later on. It's enabled by default
in OE-core because the distro might also be built with other libc
implementation, but we are not currently doing that.

Testing and main build phase must agree on the TMPDIR value. Currently
the "headless/x11/wayland" key words are spread across different build
files. It probably would be better to refactor that so that they can
be configured in refkit-ci.conf.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>