fwupd: more modular packaging
The example vendor config is not normally needed and can be separate.
Same for LVFS.

We want to use fwupd without always trusting LVFS, so keys related to
that must be packaged separately. We could put them into fwupd-lvfs,
but perhaps they also have a use for other cases (like distributing
files from LVFS via other means), so they are now in fwupd-pki.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>