fwupd-demo-firmware: package Logitech firmware
This shows how to deliver firmware updates to fwupd via the filesystem
without relying on GPG.

The .xml snipped is from LVFS and can be distributed under "CC0-1.0 OR
CC-BY-3.0+ OR CC-BY-SA-3.0+ OR GFDL-1.1+ OR FSFAP", according to
Richard Hughes. But the firmware files can only be distributed by LVFS
servers. Downloading them for a local build is fine, redistributing
them is not.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>