refkit: revise kernel reconfiguration, add TPM2
The previous approach of setting KERNEL_FEATURES globally
had some problems:
- it was based on the assumption that all BSPs have meta
  data which contains the specified features, breaking
  compilation for those BSPs which don't have them or have them
  under different names
- it couldn't be used to enable features that are not yet
  supported by the BSPs, like TPM2

To address both of these problems, kernel config fragments are stored
again together with the distro itself. The files were copied from
yocto-kernel-cache and meta-measured, so the actual config is
unchanged.

The new linux-%.bbappend is meant to be a permanent config change and
does not address a temporary issue in one of the upstream layers,
therefore it intentionally gets created under
meta-refkit-core/recipes-bsp and not somewhere under
meta-refkit-core/bbappends.

The alternative would be to leave kernel reconfiguration to the
developer who sets up a build. This would lead to a worse
out-of-the-box experience when building refkit and therefore has been
rejected.

For a discussion of this problem see
https://bugzilla.yoctoproject.org/show_bug.cgi?id=8191

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>