refkit-image.bbclass: make read-only rootfs truly read-only
OE-core does not tell the initramfs to mount read-only when
read-only-rootfs is set for an image. That looks like an oversight
and (besides touching the filesystem by mounting it read/write for
each boot) has the effect that systemd is modifying /etc/machine-id on
first boot before making the filesystem read-only.

This shows up in a "swupd verify" operation when using the active
partition as source for the inactive partition.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>