httpupdate.py: fix http server port selection
The range only included a single port, which was set this way when
debugging the code. Now it's a proper range.

The error resulting from not finding a free port was not reported
properly: the revised class has no access to the unittest fail(), so
now it just throws a RuntimeError.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>