initramfs-framework-refkit-luks.bb: additional error checking
When we failed to retrieve a password, zapping the empty keyfile led
to dd being called with bs=0, triggering an error. Using count=0 is
fine. While at it, the regular dd output gets silenced because it is
irrelevant and might even expose some undesirable information (length
of password).

Not finding a password was not reported well. Now there is an explicit
error message about it.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>